multi: allow empty activity description and validate random activity
Allow parsing of empty activity descriptions to facilitate default run
with random activity generation. We still use a vector in the
Simulation struct to allow a path where we have a combination of
random and specific activity.

This commit also adds validation that we have at least two nodes when
running random activity. This is required because destination nodes
are chosen from within the nodes that we have execution on (to prevent
liquidity draining and eventual death spirals).