sim-lib: use multiple RNGs to decrease variance of fixed seed runs
As-is, we are sharing a single RNG for all of our random operations -
destination selection for each payment, and selection of payment
parameters per-sending node. All of these operations may complete in
in different orders every time we run (be it because a task gets access
to a mutex in a different order, or a payment is faster or slower to
dispatch on the actual underlying node). This leads to large variance
in our supposedly deterministic run, because the values we get from
our "fixed" seed are different based on this order of operations.

Testing the simulator over 50 runs with one week of data generation,
the mean / std dev of the average revenue of a node = ~3.8 when ideally
we'd want this value to be < 0.1, indicating that our variance is very
high.

To address this, we use separate RNGs per-process, which reduces the
variety in order of operations impact on the values that each task gets
from its RNG. Once place where we can't address this is the network
generator, as this is shared across tasks (because it holds the whole
network graph). Selection of destination for payments will still be
subject to the randomness of our order of operations.