sim-lib: use multiple RNGs to decrease variance of fixed seed runs
As-is, we are sharing a single RNG for all of our random operations -
destination selection for each payment, and selection of payment
parameters per-sending node. All of these operations may complete in
in different orders every time we run (be it because a task gets access
to a mutex in a different order, or a payment is faster or slower to
dispatch on the actual underlying node). This leads to large variance
in our supposedly deterministic run, because the values we get from
our "fixed" seed are different based on this order of operations.

To address this, we use separate RNGs per-task, which reduces the
variety in order of operations impact on the values that each task gets
from its RNG. We can't do this for network_generator (because we use
one for all of our tasks), so we take the approach of passing a single
RNG into both our Payment and Network generator. This is admittedly
a little clunky API-wise, IMO a great improvement in the feature.

Testing this change against the original approach with:
- 50x runs of the simulator
- Sum up total revenue for a target node per run
- Look at the coefficient of variance (mean / std dev) for each data set

Results:
- Old approach: CoV = 0.5
- New approach: CoV = 40

This increase indicates a large *decrease* in the variance in the
revenue of a single node across runs (which is what we want, more
deterministic outcomes with a fixed seed).