sim-node: add a forward_only option to exclude from activity generation
Add an optional bool to simulated network channels to exclude them from
our simulation (but use them for routing in the virtual network). These
channels do not contribute to our randomly generated activity, but will
still be used to route payments.

Nodes that have entirely forward_only channels will not even be
presented to the simulation. However, there is an edge case where a
node that we want to include in the simulation is connected to a
forward_only channel. In this case, we'll exclude it from any API calls
made by the simulator, as if it isn't there.

This exclusion is a little unintuitive, because we're hiding channels
from the simulator (eg, via list_channels) because this output is used
to determine random activity volume. A more specific API could make
sense here, but that has the downside of requiring we implement it for
each node type.