simln-lib: do not listen on shutdown in produce_simulation_results
Then the Sender<SimulationOutput> is dropped by setup_payemnt_events,
the corresponding Receiver<SimulationOutput> will send None once all
of the events have been drained from it. If we listen on shutdown
triggers, we don't have the opportunity to drain this channel and
complete reporting all of our payments.

Remove listening on the listener, as we can just rely on canonical
channel mechanics.